Switch from output seeking to input seeking
Without this change, seeking inside a huge file gets slower the further you seek. E.g., if you want to have a small AudioSegment with 10 seconds starting from 6900 seconds, it may take up to 5 seconds to extract the segment.
With this change, it takes only few milliseconds!
According to:
https://trac.ffmpeg.org/wiki/Seeking, this is possible, starting from FFmpeg 2.1:
Citation:
"As of FFmpeg 2.1, when transcoding with ffmpeg (i.e. not stream copying), -ss is now also "frame-accurate" even when used as an input option"